backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
sqlmodel>=0.0.22
psycopg2-binary>=2.9.0
pgvector>=0.3.0
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0
pypdf>=4.0.0
python-docx>=1.1.0
unstructured>=0.15.0
pandas>=2.2.0
python-dotenv>=1.0.0
pydantic-settings>=2.5.0
sse-starlette>=2.1.0
